+++ app/src/main/java/pro/network/freshcatch/cart/CartItemClick.java
@@ -2,7 +2,7 @@ package pro.network.freshcatch.cart;
 
 public interface CartItemClick {
 
-    void OnQuantityChange(int position,int qty);
+    void OnQuantityChange(int position,float qty);
     void ondeleteClick(int position);
 
 }
